🚚 Rename ConditionCheckResult
to AhbExpressionEvaluationResult

and file condition_check_results.py to evaluation_results
