+++ src/ahbicht/expressions/format_constraint_expression_evaluation.py
@@ -13,7 +13,7 @@ import inject
 from lark import Token, Tree, v_args
 from lark.exceptions import VisitError
 
-from ahbicht.condition_check_results import FormatConstraintEvaluationResult
+from ahbicht.evaluation_results import FormatConstraintEvaluationResult
 from ahbicht.content_evaluation.fc_evaluators import FcEvaluator
 from ahbicht.expressions.base_transformer import BaseTransformer
 from ahbicht.expressions.condition_expression_parser import parse_condition_expression_to_tree
